fixed variable types for warnings

+++ src/main.cpp
@@ -148,7 +148,7 @@ int main() {
           vector<double> mpc_x_vals;
           vector<double> mpc_y_vals;
 
-          for (unsigned int i = 2; i < vars.size(); i = i+2) {
+          for (size_t i = 2; i < vars.size(); i = i+2) {
             mpc_x_vals.push_back(vars[i]);
             mpc_y_vals.push_back(vars[i+1]);
           }
@@ -166,7 +166,7 @@ int main() {
           //.. add (x,y) points to list here, points are in reference to the vehicle's coordinate system
           // the points in the simulator are connected by a Yellow line
 
-          for (double i = 0; i < car_ref_x.size(); i++){
+          for (unsigned int i = 0; i < car_ref_x.size(); i++){
             next_x_vals.push_back(car_ref_x(i));
             next_y_vals.push_back(car_ref_y(i));
           }
